requirements.txt
dcase-util==0.2.15
ray==0.8.7
ray[tune]==0.8.7
librosa==0.7.2
pandas==1.0.0
numpy==1.16.4
SoundFile==0.10.3.post1
tabulate==0.8.7
torchsummary==1.5.1
numba==0.48
torchvision==0.7.0
